+++ src/main/java/com/github/mlaursen/mybrews/constants/MaltType.java
@@ -1,0 +1,15 @@
+/**
+ * 
+ */
+package com.github.mlaursen.mybrews.constants;
+
+/**
+ *
+ * @author mlaursen
+ *
+ */
+public enum MaltType {
+  DRY_MALT_EXTRACT,
+  MALT_SYRUP,
+  OTHER;
+}
