+++ src/main/java/com/github/mlaursen/mybrews/entity/lookup/Malt.java
@@ -4,8 +4,11 @@
 package com.github.mlaursen.mybrews.entity.lookup;
 
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.xml.bind.annotation.XmlRootElement;
 
+import com.github.mlaursen.mybrews.constants.MaltType;
 import com.github.mlaursen.mybrews.entity.GeneratedIdNamedEntity;
 
 /**
@@ -16,5 +19,21 @@ import com.github.mlaursen.mybrews.entity.GeneratedIdNamedEntity;
 @Entity
 @XmlRootElement
 public class Malt extends GeneratedIdNamedEntity {
+  
+  @Enumerated(EnumType.STRING)
+  private MaltType type;
 
+  /**
+   * @return the type
+   */
+  public MaltType getType() {
+    return type;
+  }
+
+  /**
+   * @param type the type to set
+   */
+  public void setType(MaltType type) {
+    this.type = type;
+  }
 }
